+++ lib/lintConfig.ts
@@ -251,8 +251,11 @@ const defaultLintRuleConfig: Required<LintRuleConfig<FullLintConfigValue>> = {
 	'void-ext': [
 		2,
 		{
+			// dynamicpagelist: 2,
 			// languages: 2,
+			// rss: 2,
 			// section: 2,
+			// templatedata: 2,
 			// templatestyles: 2,
 		},
 	],