+++ src/tagPair/ext.ts
@@ -21,6 +21,8 @@ import {cached} from '../../mixin/cached';
 
 export interface ExtToken extends AttributesParentBase {}
 
+const nonVoidExt = new Set(['dynamicpagelist', 'templatedata', 'rss']);
+
 /**
  * extension tag
  *
@@ -214,9 +216,10 @@ export abstract class ExtToken extends TagPairToken {
 	override lint(start = this.getAbsoluteIndex(), re?: RegExp): LintError[] {
 		LINT: {
 			const errors = super.lint(start, re),
+				{name} = this,
 				{lintConfig} = Parser,
 				rect = new BoundingRect(this, start);
-			if (this.name !== 'nowiki') {
+			if (name !== 'nowiki') {
 				const s = this.inHtmlAttrs(),
 					rule = 'parsing-order',
 					severity = s && lintConfig.getSeverity(rule, s === 2 ? 'ext' : 'templateInTable');
@@ -224,7 +227,7 @@ export abstract class ExtToken extends TagPairToken {
 					errors.push(generateForSelf(this, rect, rule, 'ext-in-html', severity));
 				}
 			}
-			if (this.name === 'ref') {
+			if (name === 'ref') {
 				let rule: LintError.Rule = 'var-anchor',
 					s = lintConfig.getSeverity(rule, 'ref');
 				if (s && this.closest('heading-title')) {
@@ -234,6 +237,12 @@ export abstract class ExtToken extends TagPairToken {
 				s = lintConfig.getSeverity(rule, 'ref');
 				if (s && this.closest('link-text,ext-link-text')) {
 					errors.push(generateForSelf(this, rect, rule, 'ref-in-link', s));
+				}
+			} else {
+				const rule = 'void-ext',
+					s = lintConfig.getSeverity(rule, name);
+				if (s && nonVoidExt.has(name) && !this.innerText) {
+					errors.push(generateForSelf(this, {start}, rule, Parser.msg('not-empty', name), s));
 				}
 			}
 			return errors;